test(eval): consolidate ab-test command-flow + unhappy paths into one file
Merge ab-test.write.test.tsx + ab-test.create.test.tsx into a single
ab-test.test.tsx (mirrors batch-evaluation.test.tsx): hierarchy, get/list/
pause/resume/stop/delete happy paths, and every unhappy path in one place --
missing --id (now covers get, which regressed), Core-error surfacing per op
(not-found / invalid-transition / not-stopped), and config-bundle run
validation (required flags, malformed + mis-shaped JSON, identical variants,
weight bounds). Golden fixture files unchanged.
